basic demo ready
I've completed the basic shell. It features a movable player, a room
with a nifty background and a view which follows the player, and a
citizen with a random sprite.

There is also a fullscreen toggle (F11) but it's broken in the latest
version of GameMaker. Haven't found a workaround yet.