children, houses, clock
I added the child sprites, created a 'child' object that's a child of
the 'citizen' object, and made some 'block' objects like houses, with
masks (so that the player can walk behind the top part of the house,
like in old 2D RPGs).

I also reorganized the code a little. It occurs to me that we can solve
a lot of problems if we keep sections of complicated code that are
likely to need a lot of modification in separate scripts, so I created
empty scripts for CitizenMove() and PlayerAttack(), where the real code
can go later. Then, to update that code on GitHub, only the script file
needs to be uploaded, not the object file.

Player movement was broken so I changed it a little. I also crossed the
first item off the todo.txt list by implementing a countdown clock,
although nothing actually happens when it reaches 0.