Intro Screen/ Intro Narration
Rains,

Check out the narration that I have come up with so far.
I have the text scrolling across the screen and fading away after it
goes to a new line. I felt that it sort of represented the fleeting time
the player will face when they start the game.