altered AI and spawn times slightly, added more dialogue
the number of max NPCs per level now depends on the level; currently
it's 20 for each NPC (set in npcManager). the max children per level is
the max NPCs total, divided by 5.

if a spawned NPC collides with some obstacle, it'll be repositioned to
another random spot, up to 1000 times; if all those attempts fail, the
NPC will not be spawned. not a great solution, but good enough for this
game.

i also added a bunch of new dialogue for everybody.